requirements.txt
anthropic
fastapi
starlette
authlib
pydantic
passlib
gradio
google-generativeai
google-cloud-aiplatform
uvicorn
passlib
python-multipart